feat: refine directory sturucture and images based on review
