+++ frontend/src/app/(public)/layout.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+export default function PublicLayout({
+  children,
+}: Readonly<{
+  children: React.ReactNode;
+}>) {
+  return <>{children}</>;
+}
