+++ frontend/src/components/layouts/Footer.tsx
@@ -1,4 +1,3 @@
-import linkedInLogo from "@/public/images/LI-In-Bug.png";
 import Image from "next/image";
 import Link from "next/link";
 
@@ -24,28 +23,28 @@ export default function Footer() {
               href="https://www.linkedin.com/in/risa-yamamoto-profile/"
               className="flex items-center gap-1 hover:opacity-70"
             >
-              <Image src={linkedInLogo} alt="LinkedIn" width={20} height={20} />
+              <Image src="/images/LI-In-Bug.png" alt="LinkedIn" width={20} height={20} />
               <p className="text-text text-sm">Risa</p>
             </Link>
             <Link
               href="https://www.linkedin.com/in/ben-ono-profile/"
               className="flex items-center gap-1 hover:opacity-70"
             >
-              <Image src={linkedInLogo} alt="LinkedIn" width={20} height={20} />
+              <Image src="/images/LI-In-Bug.png" alt="LinkedIn" width={20} height={20} />
               <p className="text-text text-sm">Ben</p>
             </Link>
             <Link
               href="http://linkedin.com/in/nao-g-55822a2bb"
               className="flex items-center gap-1 hover:opacity-70"
             >
-              <Image src={linkedInLogo} alt="LinkedIn" width={20} height={20} />
+              <Image src="/images/LI-In-Bug.png" alt="LinkedIn" width={20} height={20} />
               <p className="text-text text-sm">Nao</p>
             </Link>
             <Link
               href="https://www.linkedin.com/in/haruka-kakiuchi-profile/"
               className="flex items-center gap-1 hover:opacity-70"
             >
-              <Image src={linkedInLogo} alt="LinkedIn" width={20} height={20} />
+              <Image src="/images/LI-In-Bug.png" alt="LinkedIn" width={20} height={20} />
               <p className="text-text text-sm">Haruka</p>
             </Link>
           </div>
